% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{permeated_at}
\alias{permeated_at}
\title{Permeated mass per unit area at specific time(s)}
\usage{
permeated_at(res, t)
}
\arguments{
\item{res}{A `skin_result` object.}

\item{t}{Time(s) at which to evaluate. Must have units of time.}
}
\value{
A units vector of permeated amounts in mass-per-area.
}
\description{
Linearly interpolates the cumulative permeated curve [permeated()] to
the requested time(s).
}
