% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skindiff-package.R
\docType{package}
\name{skindiff-package}
\alias{skindiff}
\alias{skindiff-package}
\title{skindiff: numerical simulation of drug diffusion through skin}
\description{
Builds a 1-D Crank-Nicolson finite-difference model of a stack of skin
compartments (vehicle / stratum corneum / deeper skin layers / sink) and
returns time-resolved mass and concentration-depth profiles.
}
\author{
\strong{Maintainer}: Dominik Selzer \email{d.selzer@scientific-consilience.com}

}
\keyword{internal}
